Refactoring sample data
