+++ db/sample_data/agreement_letters.rb
@@ -1,0 +1,7 @@
+def agreement_letter_applicant_gongakrobatik(user, event)
+  AgreementLetter.new(
+    user: user,
+    event: event,
+    path: "/storage/agreement_letters/foo.pdf"
+  )
+end
