+++ db/sample_data/requests.rb
@@ -1,0 +1,6 @@
+def request_hardware_entwicklung(user)
+  Request.new(
+    topics: "Hardware-Entwicklung mit einem CAD-System",
+    user: user
+  )
+end
